Convert publish static pages unit test
